Rename apiFactory file from utils.ts to apiFactory.ts
