+++ src/api/registries/mutate/deleteRegistry.ts
@@ -5,7 +5,7 @@ import {
 } from "../schema";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 
-import { apiFactory } from "../../utils";
+import { apiFactory } from "../../apiFactory";
 import { registriesKeys } from "..";
 import { useApiKey } from "../../../util/store/apiKey";
 import { useNamespace } from "../../../util/store/namespace";
