+++ src/api/tree/mutate/createDirectory.ts
@@ -1,5 +1,5 @@
 import { FolderCreatedSchema } from "../schema";
-import { apiFactory } from "~/api/utils";
+import { apiFactory } from "~/api/apiFactory";
 import { forceLeadingSlash } from "../utils";
 import { useApiKey } from "~/util/store/apiKey";
 import { useMutation } from "@tanstack/react-query";
