+++ src/api/tree/mutate/createRevision.tsx
@@ -1,7 +1,7 @@
 import { ToastAction, useToast } from "~/design/Toast";
 
 import { WorkflowCreatedSchema } from "../schema";
-import { apiFactory } from "~/api/utils";
+import { apiFactory } from "~/api/apiFactory";
 import { forceLeadingSlash } from "../utils";
 import { pages } from "~/util/router/pages";
 import { useApiKey } from "~/util/store/apiKey";
