+++ src/api/tree/mutate/updateWorkflow.ts
@@ -1,7 +1,7 @@
 import { NodeListSchemaType, WorkflowCreatedSchema } from "../schema";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 
-import { apiFactory } from "~/api/utils";
+import { apiFactory } from "~/api/apiFactory";
 import { forceLeadingSlash } from "../utils";
 import { treeKeys } from "..";
 import { useApiKey } from "~/util/store/apiKey";
