+++ src/api/tree/query/node.ts
@@ -2,7 +2,7 @@ import { forceLeadingSlash, sortFoldersFirst } from "../utils";
 
 import { NodeListSchema } from "../schema";
 import type { QueryFunctionContext } from "@tanstack/react-query";
-import { apiFactory } from "~/api/utils";
+import { apiFactory } from "~/api/apiFactory";
 import { treeKeys } from "..";
 import { useApiKey } from "~/util/store/apiKey";
 import { useNamespace } from "~/util/store/namespace";
