+++ src/api/tree/query/router.ts
@@ -2,7 +2,7 @@ import { forceLeadingSlash, sortByRef } from "../utils";
 
 import type { QueryFunctionContext } from "@tanstack/react-query";
 import { RouterSchema } from "../schema";
-import { apiFactory } from "../../utils";
+import { apiFactory } from "../../apiFactory";
 import { treeKeys } from "../";
 import { useApiKey } from "../../../util/store/apiKey";
 import { useNamespace } from "../../../util/store/namespace";
